pyproject.toml
[project]
name = "tg-emoji-casino"
version = "0.1.0"
description = "Telegram emoji casino bot (aiogram + FastAPI)"
authors = [{name = "Team TG Emoji Casino", email = "dev@gnezdoai.ru"}]
requires-python = ">=3.11"
dependencies = [
    "aiogram==3.6.0",
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "SQLAlchemy==2.0.30",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "redis==5.0.4",
    "pydantic-settings==2.2.1",
    "python-dotenv==1.0.1",
    "prometheus-client==0.20.0",
    "PyJWT==2.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.1",
    "pytest-asyncio==0.23.7",
    "httpx==0.27.0",
    "aiosqlite==0.20.0"
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*"]
